[proposa] Streamline, simplify README
I felt overwhelmed and confused reading the README and trying to get started. There were too many pieces, too many options. I wrote this as an example how it could be simplified so that a newcomer can find her way to running and interacting with the app more quickly.

It would be even better if `clj -A:dev ...` actually started a background nREPL server in addition to the interactive REPL itself, since you could then connect to it from Cursive/... and use the Quick Start as-is to start coding, similarly to what `lein repl` does.